ShopperStack_Framework/requirements.txt
requests==2.32.2
pytest==8.3.3
pytest-html==4.1.1
python-dotenv==1.0.1